// src/app/models/enums/law-subject-matter-example.enum.ts
export enum LawSubjectMatterExampleEnum {
  //Civil
  SMALL_CLAIMS = 'Small Claims',
  EMPLOYMENT_CLAIMS = 'Employment Claims',
  NEIGHBOUR_DISPUTES = 'Neighbour disputes',
  BANKRUPTCY = 'Bankruptcy',
  BREACH_OF_CONTRACT = 'Breach of Contract',
  COMMERCIAL = 'Commercial disputes',
  //Criminal
  HOMICIDE = 'Homicide',
  DRUG_OFFENCES = 'Drug offences',
  SEXUAL_OFFENCES = 'Sexual offences',
  CHEATING = 'Cheating',
  TRAFFIC_OR_REGULATORY = 'Traffic or regulatory offences',
  //Family
  DIVORCE = 'Divorce',
  FAMILY_VIOLENCE = 'Family violence',
  PROBATE_AND_ADMIN = 'Probate and administration',
  MAINTENANCE = 'Maintenance',
  ADOPTION = 'Adoption',
  CHILD_PROTECTION = 'Child Protection',

}